backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "post-meeting-automation-backend"
version = "0.1.0"
description = "Backend for post-meeting task extraction and delivery."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "google-auth>=2.38.0,<3.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "psycopg[binary]>=3.2.3,<4.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
